#86 - Rename: 잘못된 페이지 표기법 수정 (Feed -> Main)
